LayoutTests/fast/js/typedarray-set-destination-smaller-than-source-expected.txt

Tests the code path in typedArray.set that may have to do a copy via an intermediate buffer because the source and destination overlap and have different size elements (source is larger than destination).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo(64) is foo_reference(64)
PASS bar(64) is bar_reference(64)
PASS baz(64) is baz_reference(64)
PASS successfullyParsed is true

TEST COMPLETE
